Implement client_util
